+++ sparkle_log/scheduler.py
@@ -5,13 +5,19 @@ from threading import Event
 
 import schedule
 
+from sparkle_log.custom_types import CustomMetricsCallBacks, GraphStyle, Metrics
 from sparkle_log.log_writer import log_system_metrics
-from sparkle_log.ui import GraphStyle
 
 
-def run_scheduler(stop_event: Event, metrics: tuple[str, ...], seconds: int, style: GraphStyle = "bar"):
+def run_scheduler(
+    stop_event: Event,
+    metrics: tuple[Metrics, ...],
+    seconds: int,
+    style: GraphStyle = "bar",
+    custom_metrics: CustomMetricsCallBacks = None,
+):
     """Run scheduled tasks until the stop_event is set."""
-    schedule.every(seconds).seconds.do(lambda: log_system_metrics(metrics, style))
+    schedule.every(seconds).seconds.do(lambda: log_system_metrics(metrics, style, custom_metrics))
 
     while not stop_event.is_set():
         schedule.run_pending()