+++ sparkle_log/ui.py
@@ -2,14 +2,14 @@
 Sparkline graph code goes here.
 """
 
-from typing import Literal
+from typing import cast
 
 import sparklines
 
-GraphStyle = Literal["bar", "faces", "jagged", "linear", "vertical", "ascii_art", "pie_chart"]
+from sparkle_log.custom_types import GraphStyle, NumberType
 
 
-def sparkline(numbers: list[int], style: GraphStyle = "bar") -> str:
+def sparkline(numbers: list[NumberType], style: GraphStyle = "bar") -> str:
     """Generate a simple sparkline string for a list of integers."""
     if style == "bar":
         for line in sparklines.sparklines(numbers):
@@ -29,20 +29,20 @@ def sparkline(numbers: list[int], style: GraphStyle = "bar") -> str:
     return ""
 
 
-def faces_sparkline(data):
+def faces_sparkline(data: list[NumberType]):
     """Generate a sparkline with face emojis."""
     symbols = ["😞", "😐", "😊", "😁"]
     return sparkline_it(data, symbols)
 
 
-def sparkline_it(data: list[int], symbols: list[str]):
+def sparkline_it(data: list[NumberType], symbols: list[str]):
     """Generate a sparkline with the given symbols."""
     noneless_data = [_ for _ in data if _ is not None]
     max_val = max(noneless_data)
     min_val = min(noneless_data)
     range_val = max_val - min_val
     if range_val == 0:  # Avoid division by zero
-        return "".join(symbols[0] for _ in data)
+        return "".join(" " for _ in data)
     if range_val * len(symbols) == 0:
         return " "
     return "".join(
@@ -51,31 +51,40 @@ def sparkline_it(data: list[int], symbols: list[str]):
     )
 
 
-def pie_chart_sparkline(data):
+def pie_chart_sparkline(data: list[NumberType]):
     """Using different geometric shapes or other symbols"""
     symbols = ["○", "◔", "◑", "◕", "●"]
     return sparkline_it(data, symbols)
 
 
-def ascii_sparkline(data):
+def ascii_sparkline(data: list[NumberType]):
     """Using different ASCII characters."""
     symbols = [" ", ".", ":", "-", "=", "+", "*", "#", "%", "@"]
     return sparkline_it(data, symbols)
 
 
-def linear_ascii_sparkline(data):
+def linear_ascii_sparkline(data: list[NumberType]):
     """Ascii for Low, medium, and high levels"""
     levels = ["_", "-", "¯"]
     return sparkline_it(data, levels)
 
 
-def jagged_ascii_sparkline(data):
+def jagged_ascii_sparkline(data: list[NumberType]):
     """Ascii incorporating a peak character"""
     symbols = ["_", "-", "^", "¯"]
     return sparkline_it(data, symbols)
 
 
-def vertical_ascii_sparkline(data):
+def vertical_ascii_sparkline(data: list[NumberType]):
     """Ascii Using single and double vertical lines"""
     symbols = ["_", "|", "‖"]
     return sparkline_it(data, symbols)
+
+
+if __name__ == "__main__":
+
+    def run():
+        for style in ["bar", "jagged", "vertical", "linear", "ascii_art", "pie_chart", "faces"]:
+            print(f"{style}: {sparkline([1, 2, 3, 4, 5, 6, 7, 8, 9, 10], cast(GraphStyle, style))}")
+
+    run()