+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.4.0] - 2024-04-27
+
+### Added
+
+- Decorator support for async functions
+- Start support for arbitrary custom metrics
+
+### Fixed
+- Bug on new graph types when denominator would be 0. 
+
 ## [0.3.0] - 2024-04-27
 
 ### Added